doesnt work :|

+++ 08/TENMAJKL.php
@@ -1,9 +1,12 @@
 <?php
 
-function fmap(int $x, int $a, int $b, int $c, int $d): float
+// TODO
+
+function fmap(float $x, float $a, float $b, float $c, float $d): float
 {
     $shift = $a - $c;
     $b_shifted = $b - $shift;
     $mult = $d / $b_shifted;
     return ($x - $shift) * $mult;
 }
+
